+++ NEWS.md
@@ -1,5 +1,6 @@
 # rfp (development version)
 
+  - convert inputs to `rfp_bcd_get_data` to uppercase to allow more flexible input.
   - add `rfp_source_csv.R`
   - add `rfp_source_csv.R`
   - add `rfp_u_rm_gpkg_layers.R`